requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1
idna==3.6
redis==5.0.1
requests==2.31.0
six==1.16.0
spotipy==2.23.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.2.0
celery==5.3.4
django-celery-results==2.5.1
kombu==5.3.4
cryptography==42.0.4
python-decouple==3.8
pillow==10.1.0
lyricsgenius==3.0.1
